requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==2.5.2  # Versione aggiornata
pydantic-settings==2.1.0  # Nuova dipendenza
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.0.0